+++ modules/core/src/test/java/org/apache/ignite/internal/GridTaskFutureImplStopGridSelfTest.java
@@ -182,7 +182,7 @@ public class GridTaskFutureImplStopGridSelfTest extends GridCommonAbstractTest {
             int res = 0;
 
             for (ComputeJobResult result : results) {
-                res += (Integer)result.getData();
+                res += result.getData();
             }
 
             return res;
