+++ modules/core/src/test/java/org/apache/ignite/session/GridSessionFutureWaitTaskAttributeSelfTest.java
@@ -256,7 +256,7 @@ public class GridSessionFutureWaitTaskAttributeSelfTest extends GridCommonAbstra
 
             for (ComputeJobResult result : results) {
                 if (result.getData() != null)
-                    sum += (Integer)result.getData();
+                    sum += result.getData();
             }
 
             return sum;
